requirements.txt
aiogram==3.12.0
python-dotenv==1.0.1
yt-dlp==2025.9.21.232818.dev0
aiohttp==3.9.1
Pillow==10.1.0
ffmpeg-python==0.2.1
requests==2.31.0
